requirements.txt
# Django
Django==3.2
psycopg2==2.8.6

# Rest Framework
requests==2.23.0
djangorestframework==3.12.0
django-oauth-toolkit==1.3.2
drf-yasg==1.20.0

# GraphQL
graphene-django==2.15.0
django-graphql-auth==0.3.16
